+++ src/Server/Game/enemies/walkingEnemy.h
@@ -13,7 +13,7 @@
 class WalkingEnemy: public Enemy {
 
 public:
-    WalkingEnemy(GameMap& map, Vector<int16_t> pos, int16_t id);
+    WalkingEnemy(GameMap& map, Vector<uint8_t> pos, uint8_t id);
 
     EnemyType getEnemyType() override;
 };
