+++ src/Server/Game/weapons/bouncer.h
@@ -11,19 +11,19 @@
 
 class Bouncer: public Weapon {
 private:
-    uint16_t bullets;
-    const uint16_t maxBullets;
-    const uint16_t damage;
+    uint8_t bullets;
+    const uint8_t maxBullets;
+    const uint8_t damage;
     const float fireRate;
-    uint32_t bulletsShot = 0;
-    int16_t lastTimeShot;
+    uint8_t bulletsShot = 0;
+    uint8_t lastTimeShot;
 
 public:
     Bouncer();
 
     void update(float time) override;
 
-    void shoot(std::vector<std::shared_ptr<Entity>>& shootingEntities, int16_t xPos,
+    void shoot(std::vector<std::shared_ptr<Entity>>& shootingEntities, uint8_t xPos,
                float time) override;
 
     void reload() override;
@@ -32,9 +32,9 @@ public:
 
     bool isLoaded() override;
 
-    int16_t getBullets() override;
+    uint8_t getBullets() override;
 
-    uint32_t getBulletsShot() override;
+    uint8_t getBulletsShot() override;
 
     bool shootTime(float time) override;
